fix bag with int

+++ f.cpp
@@ -7,9 +7,9 @@ unsigned long long fib(int n){
   if (n < 3){
     return 1;
   }
-  int n1 =  1;
-  int n2 =  2;
-  int temp = 0;
+unsigned long long n1 =  1;
+ unsigned long long n2 =  2;
+  unsigned long long temp = 0;
   for (int i =3 ; i <= n; i++){
 temp = n1 +n2;
   n2 = n1;
